refactor(weather): correct names for layers
