+++ services/weather/src/cache/infrastructure/in-memory-cache.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { CacheService } from './cache.service';
+import { CacheService } from '../domain/cache.service';
 
 @Injectable()
 export class InMemoryCacheService implements CacheService {
