+++ services/weather/src/weather/infrastructure/providers/helpers/weather-url.builder.ts
