+++ services/weather/src/weather/infrastructure/providers/openweathermap.provider.ts
@@ -5,7 +5,7 @@ import { HttpService } from '@nestjs/axios';
 import { AxiosResponse } from 'axios';
 import { lastValueFrom } from 'rxjs';
 import { OpenWeatherMapResponse } from './types/openweathermap-response.type';
-import { buildOpenWeatherMapUrl } from '../helpers/weather-url.builder';
+import { buildOpenWeatherMapUrl } from './helpers/weather-url.builder';
 import { WeatherLogger } from '../logger/weather.logger';
 
 @Injectable()
@@ -27,8 +27,10 @@ export class OpenWeatherMapProvider implements WeatherProvider {
     );
 
     try {
-      const response: AxiosResponse<OpenWeatherMapResponse> =
-        await lastValueFrom(this.httpService.get<OpenWeatherMapResponse>(url));
+      // eslint-disable-next-line @typescript-eslint/no-unsafe-call
+      const response = (await lastValueFrom(
+        this.httpService.get<OpenWeatherMapResponse>(url),
+      )) as AxiosResponse<OpenWeatherMapResponse>;
 
       // api error imitation
       // throw new Error('Server unavailable');
